fix: update OpenAI embedding prediction to handle single and multiple inputs
- Modified `predict` method in `TestEmbedAPI` to support both single and batch inputs by adjusting the random embedding generation.
- Added comprehensive tests for OpenAI embedding spec, covering single input, multiple inputs, usage validation, and error handling for missing or incorrect responses.
- Ensured that the tests validate the expected structure and content of the API responses.
